% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_catalog}
\alias{fetch_catalog}
\title{Fetch the Copernicus catalog (all of it)}
\usage{
fetch_catalog(
  filename = nicolaus_path("catalogs", "all_products_copernicus_marine_service.json"),
  verbose = FALSE,
  app = copernicus::get_copernicus_app()
)
}
\arguments{
\item{filename}{chr, the output filename}

\item{verbose}{logical, if TRUE issue progress messages}

\item{app}{chr, the copernicusmarine application}
}
\value{
0 (success) or non-0 (failure)
}
\description{
Fetch the Copernicus catalog (all of it)
}
