% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{flatten_dataset}
\alias{flatten_dataset}
\title{Flatten one or more dataset nodes}
\usage{
flatten_dataset(
  dd = example_datasets(),
  service = "arco-geo-series",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{dd}{a "datasets" list}

\item{service}{chr the preferred service}

\item{verbose}{logical, output messages if TRUE}

\item{...}{arguments passed to flatten_variables}
}
\value{
table of product information
}
\description{
Flatten one or more dataset nodes
}
