% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product.R
\name{flatten_product}
\alias{flatten_product}
\title{Flatten one or more product nodes node}
\usage{
flatten_product(pp = example_products(), verbose = FALSE, ...)
}
\arguments{
\item{pp}{a "products" list of one or more "product" objects}

\item{verbose}{logical, output messages if TRUE}

\item{...}{arguments passed to flatten_dataset}
}
\value{
table of product information
}
\description{
Flatten one or more product nodes node
}
