% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{flatten_variable}
\alias{flatten_variable}
\title{Flatten variable(s)}
\usage{
flatten_variable(vv = example_variables())
}
\arguments{
\item{vv}{list of one or more variables}
}
\value{
table of flattened variable(s)
}
\description{
Flatten variable(s)
}
