% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_catalog}
\alias{import_catalog}
\title{Import the json product catalog}
\usage{
import_catalog(
  filename = nicolaus_path("catalogs", "all_products_copernicus_marine_service.json"),
  flatten = TRUE
)
}
\arguments{
\item{filename}{chr, the name of the catalog name}

\item{flatten}{logical, if TRUE then transform the out to a flat table}
}
\value{
a named list of json elements or a table
}
\description{
Import the json product catalog
}
