% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicolaus-package.R
\docType{package}
\name{nicolaus-package}
\alias{nicolaus-package}
\alias{nicolaus}
\title{nicolaus: Unpacking the Copernicus Marine Data Store catalog}
\description{
Provides R-language tools to unpack the Copernicus Marine Data Store catalog.
}
\keyword{internal}
