% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{nicolaus_path}
\alias{nicolaus_path}
\title{Build a nicolaus path}
\usage{
nicolaus_path(..., root = root_path())
}
\arguments{
\item{...}{further arguments for \code{file.path()}}

\item{root}{the root path}
}
\value{
character path description
}
\description{
Build a nicolaus path
}
