% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{parse_time}
\alias{parse_time}
\title{Parse a time value}
\usage{
parse_time(x, unit = c("ISO8601"), origin = time_origin())
}
\arguments{
\item{x}{the object to parse with one or more elements}

\item{unit}{chr, provided in the time coordinate list element}

\item{origin}{POSIXct, only used if time is in ms}
}
\value{
POSIXct time value(s)
}
\description{
So far we have encountered ms since 1970-01-01 and ISO8601
date-time stamps
}
