% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_catalog}
\alias{read_catalog}
\title{Read a tabular catalog}
\usage{
read_catalog(
  filename = nicolaus_path("catalogs", "all_products_copernicus_marine_service.rds")
)
}
\arguments{
\item{filename}{chr the name of the inout file}
}
\value{
the input table
}
\description{
Read a tabular catalog
}
