% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{variable_template}
\alias{variable_template}
\title{Generate a template for variables}
\usage{
variable_template(n = 1, dataset_id = NA_character_)
}
\arguments{
\item{n}{num, the number of rows to return}

\item{dataset_id}{chr, the dataset_id (by default NA)}
}
\value{
n-row table
}
\description{
Generate a template for variables
}
