% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_catalog}
\alias{write_catalog}
\title{Save a tabular catalog}
\usage{
write_catalog(
  x,
  filename = nicolaus_path("catalogs", "all_products_copernicus_marine_service.rds")
)
}
\arguments{
\item{x}{catalog table}

\item{filename}{chr the name of the output file}
}
\value{
the input table
}
\description{
Save a tabular catalog
}
